+++ changelog.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.27.0] 2024-12-19
 
 - Add func apidoc.DocumentRouter
 - Use type apidoc.Doc to index routes for additional documentation